카카오 로그인 수정중

+++ src/main/java/com/crossroads/app/domain/vo/MemberVO.java
@@ -19,7 +19,7 @@ public class MemberVO {
     private boolean memberFileType;
     private String memberType;
     private String memberDriveRegisterDate;
-    private Long memberStatus;
+    private Integer memberStatus;
     private Integer memberPoint;
     private Long memberRandomKey;
